[ZkTracer] update `mul` and `mod` modules (#1942)
* update tracing for mul/mod modules

This updates the mul and mod packages responsible for tracing the respective
modules.  Mostly this amounts to trimming these modules down and removing
unnecessary stuff.

* update go-corset to v1.2.0-rc3

This is required to compile the constraints.

* tweak timeouts / max forks

This puts through some minor tweaks for running the tracer tests.

* update linea-constraints

This updates linea-constraints to include some minor tweaks after
review.

* remove unused stamp parameters

This removes some unsused stamp parameters.

* update linea-constraints

This updates the linea-constaints to point back to master (as this now
contains the merged MOD/MUL modules).